moved db into new extensions file and implemented basic celery setup into Flask App - not being used right now
