+++ app/__init__.py
@@ -1,18 +1,35 @@
 import os
 from flask import Flask, render_template
 from flask_assets import Environment
-from flask_sqlalchemy import SQLAlchemy
 from flask_login import LoginManager
 from flask_bootstrap import Bootstrap
 from flask_jsglue import JSGlue
+from celery import Celery
 
 from .util.config import DevConfig
+from .util.extensions import db
 
-db = SQLAlchemy()
+def create_celery_app(app=None):
+    app = app or create_app()
+    celery = Celery(__name__, broker=app.config['CELERY_BROKER_URL'])
+    celery.conf.update(app.config)
+    TaskBase = celery.Task
+
+    class ContextTask(TaskBase):
+        abstract = True
+
+        def __call__(self, *args, **kwargs):
+            with app.app_context():
+                return TaskBase.__call__(self, *args, **kwargs)
+
+    celery.Task = ContextTask
+    celery.app = app
+    return celery
 
 def create_app(config_object=DevConfig):
     app = Flask(__name__)
     app.config.from_object(config_object)
+    #celery = create_celery_app(app)
     
     register_assets(app)
     register_extensions(app)
