+++ app/views/queries.py
@@ -1,8 +1,8 @@
 from flask import Blueprint, render_template, request, redirect, url_for, jsonify
 from flask_login import current_user, login_required
 
-from .. import db
 from ..models.model import Projects, Experiments
+from ..util.extensions import db
 from ..util.table_queries import ProjectsDataTable, ExperimentsDataTable
